src/main/java/com/meretskiy/mytodover/services/impl/DefaultUserService.java
package com.meretskiy.mytodover.services.impl;

import com.meretskiy.mytodover.dto.UserDto;
import com.meretskiy.mytodover.model.Role;
import com.meretskiy.mytodover.model.User;
import com.meretskiy.mytodover.repositories.UserRepository;
import com.meretskiy.mytodover.services.RoleService;
import com.meretskiy.mytodover.services.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Collection;
import java.util.Optional;
import java.util.stream.Collectors;

@Service
@RequiredArgsConstructor
public class DefaultUserService implements UserService {

    private final UserRepository userRepository;
    private final BCryptPasswordEncoder encoder;
    private final RoleService roleService;

    @Override
    public Optional<User> findByUsername(String username) {
        return userRepository.findByUsername(username);
    }

    @Override
    @Transactional
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException(String.format("User '%s' not found", username)));
        return new org.springframework.security.core.userdetails.User(
                user.getUsername(), user.getPassword(), mapRolesToAuthorities(user.getRoles()));
    }

    private Collection<? extends GrantedAuthority> mapRolesToAuthorities(Collection<Role> roles) {
        return roles.stream().map(role -> new SimpleGrantedAuthority(role.getName())).collect(Collectors.toList());
    }

    @Override
    public User createNewUser(UserDto userDto) {
        userDto.setPassword(encoder.encode(userDto.getPassword()));
        User newUser = new User(userDto);
        newUser.getRoles().add(roleService.getRoleForNewUser());
        userRepository.save(newUser);
        return newUser;
    }
}
